requirements.txt
fastapi[standard]==0.115.*
uvicorn[standard]==0.34.*
twilio==9.*
openai==1.*
websockets==13.*
python-dotenv==1.*
pydantic-settings==2.*
chromadb>=0.5
PyPDF2>=3.0
python-docx>=1.0
audioop-lts>=0.2.1; python_version >= "3.13"
httpx>=0.27

